pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "comfyui-api-yoboxapp-node"
version = "1.0.0"
description = "AiYobox API integration node for ComfyUI"
authors = [
    {name = "AiYobox", email = "support@yoboxapp.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "requests",
    "torch",
    "numpy", 
    "Pillow"
]
keywords = ["comfyui", "ai", "image-generation", "aiyobox"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

[project.urls]
Homepage = "https://api.yoboxapp.com"
Repository = "https://github.com/yourusername/comfyui-api-yoboxapp-node"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]

[project.entry-points."comfyui.custom_nodes"]
aiyobox = "api_yoboxapp_node"